requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.7
beautifulsoup4==4.9.3
black==21.5b0
certifi==2019.6.16
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.8
django-admin-autocomplete-filter==0.6.1
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==7.1
# django-endless-pagination==2.0
django-extensions==3.1.3
django-guardian==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-nose==1.4.7
django-notifications-hq==1.6.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
humanfriendly==9.1
idna==2.8
jsonfield==3.1.0
mypy-extensions==0.4.3
nose==1.3.7
oauthlib==3.1.0
pathspec==0.8.1
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyreadline==2.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.1
stripe==2.56.0
swapper==1.1.2.post1
toml==0.10.2
urllib3==1.25.3
uuid==1.30
# waitress==2.0.0
whitenoise==5.2.0
wrapt==1.11.2
yapf==0.31.0
